Fix of LDAP password not asked after username immediately
Moved operations like adding into own module
Lib source file was packed with a lot of functions before

Adjusted Changelog
